machine: fix "OUT PINS 0" and "OUT PINDIRS 0"
According to the rp2040 datasheet 0 has a special meaning, so
that all 32 GPIO bits can be configured or set.

3.4.5.2 section:

  "Bit count: how many bits to shift out of the OSR. 1…32 bits,
   32 is encoded as 00000"

The 'pins_out_count' has 6 bits width, so 32 can be encoded for
this configuration variable as well.

Patch fixes the "OUT PINS 0" and "OUT PINDIRS 0" instructions
and makes possible to address 32 GPIO pins.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>