machine: make GPIO output delay equal to 1 cycle instead of 2
Patch removes one extra FF register for the GPIO output and
makes 'output_pins' and 'pins_directions' assignment part
of the combinational logic in the machine.v. The pio.v will
collect all the GPIO signals and push them to the FF register,
which will be visible on the next cycle.

GPIO 1 cycle delay is described in the original rp2040 pdf
datasheet in the "3.5.6.1. Output Priority" section:

  "There is a register stage between each state machine and
   the pin mapping logic ..."

   and

   "Assuming zero propagation delay, a state machine observing
    its own outputs is subject to the following delays:

     o when bypassing synchronisers, a two-cycle delay ..."

So two-cycle delay is only possible if input and output have
one FF register each.

Currently there is no any register stage or even synchronizers
between GPIO input and state machine. This patch will follow,
stay tuned.

Also this change moves the reset of the 'gpio_out' and
'gpio_dirs' to the same always block where these regsiters
are set. This fixes the following synthesis warning:

   Warning: multiple conflicting drivers for top.\pio_1.gpio_dirs [31]:
   Warning: multiple conflicting drivers for top.\pio_1.gpio_out [31]:

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>