pio: implement input GPIO register and 2-FF synchronizer
The patch implements register and 2-FF synchronizer for 0..31
input GPIO pins as it is described in the "3.7. List of Registers"
section of the rp2040 datasheet:

  PIO:INPUT_SYNC_BYPASS Register
  ------------------------------
     There is a 2-flipflop synchronizer on each GPIO input, which protects PIO logic
     from metastabilities. This increases input delay, and for fast synchronous IO
     (e.g. SPI) these synchronizers may need to be bypassed. Each bit in this
     register corresponds to one GPIO.
     0 -> input is synchronized (default)
	1 -> synchronizer is bypassed
	If in doubt, leave this register as all zeroes.

Paragraph "3.5.6.3. Input Synchronizers" covers the topic in more details.
For those who are worried about delays it is worth to read the "3.5.6.1.
Output Priority" paragraph.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>