machine,pc: fix execution of the immediate instruction
In the rp2040 datasheet, the "3.5.7. Forced and EXEC’d Instructions"
paragraph stated the following:

  ```
  When an instruction is written to the INSTR register, the state machine immediately decodes and executes that
  instruction, rather than the instruction it would have fetched from the PIO’s instruction memory. The program counter
  does not advance, so on the next cycle (assuming the instruction forced into the INSTR interface did not stall) the state
  machine continues to execute its current program from the point where it left off, unless the written instruction itself
  manipulated PC.

  Delay cycles are ignored on instructions written to the INSTR register, and execute immediately, ignoring the state
  machine clock divider. This interface is provided for performing initial setup and effecting control flow changes, so it
  executes instructions in a timely manner, no matter how the state machine is configured.
  ```

So immediate instruction can be executed in any SM state, even when
SM is disabled (this is easy to check on the real hardware and, yes,
this works).

The desired behavior is to execute immediate instruction in any state
and to keep the original program counter for any instruction, but
one which modifies the PC (e.g. JMP).

This patch fixes the original code of the PC module, making possible
execution of any immediate instruction even `penable` is cleared.
Also this change allows PC to be modified by the instruction,
which lets executing JMP instruction before SM is enabled (this is
required for setting up N SM for the same PIO).

Also patch makes combination logic part of the always block instead
of long chain in the continuous assignment, which is, to be frank,
not easy to read.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>